render search result
